+++ lib/stdnoreturn.dart
@@ -1,0 +1,9 @@
+// ignore_for_file: camel_case_types
+
+/// `<stdnoreturn.h>` implementation for stdc
+/// 
+/// Provides the `noreturn` type definition, mapping to Dart's `Never`.
+library;
+
+/// Represents a function that never returns.
+typedef noreturn = Never;